@@ -1,6 +1,6 @@
 {
   "name": "mongodb",
-  "version": "6.18.0",
+  "version": "6.19.0",
   "description": "The official MongoDB driver for Node.js",
   "main": "lib/index.js",
   "files": [
@@ -35,7 +35,7 @@
     "gcp-metadata": "^5.2.0",
     "kerberos": "^2.0.1",
     "mongodb-client-encryption": ">=6.0.0 <7",
-    "snappy": "^7.2.2",
+    "snappy": "^7.3.2",
     "socks": "^2.7.1"
   },
   "peerDependenciesMeta": {
@@ -62,15 +62,15 @@
     }
   },
   "devDependencies": {
-    "@aws-sdk/credential-providers": "^3.632.0",
+    "@aws-sdk/credential-providers": "^3.864.0",
     "@iarna/toml": "^2.2.5",
     "@istanbuljs/nyc-config-typescript": "^1.0.2",
-    "@microsoft/api-extractor": "^7.52.5",
+    "@microsoft/api-extractor": "^7.52.10",
     "@microsoft/tsdoc-config": "^0.17.1",
     "@mongodb-js/zstd": "^2.0.1",
     "@types/chai": "^4.3.17",
     "@types/chai-subset": "^1.3.5",
-    "@types/express": "^5.0.1",
+    "@types/express": "^5.0.3",
     "@types/kerberos": "^1.1.5",
     "@types/mocha": "^10.0.9",
     "@types/node": "^22.15.3",
@@ -79,35 +79,35 @@
     "@types/sinon": "^17.0.4",
     "@types/sinon-chai": "^4.0.0",
     "@types/whatwg-url": "^13.0.0",
-    "@typescript-eslint/eslint-plugin": "^8.31.1",
+    "@typescript-eslint/eslint-plugin": "^8.40.0",
     "@typescript-eslint/parser": "^8.31.1",
     "chai": "^4.4.1",
     "chai-subset": "^1.6.0",
     "chalk": "^4.1.2",
-    "eslint": "^9.25.1",
-    "eslint-config-prettier": "^10.1.2",
+    "eslint": "^9.33.0",
+    "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-mocha": "^10.4.1",
-    "eslint-plugin-prettier": "^5.2.3",
+    "eslint-plugin-prettier": "^5.5.4",
     "eslint-plugin-simple-import-sort": "^12.1.1",
     "eslint-plugin-tsdoc": "^0.4.0",
-    "eslint-plugin-unused-imports": "^4.1.4",
+    "eslint-plugin-unused-imports": "^4.2.0",
     "express": "^5.1.0",
     "gcp-metadata": "^5.3.0",
     "js-yaml": "^4.1.0",
-    "mocha": "^10.8.2",
+    "mocha": "^11.7.1",
     "mocha-sinon": "^2.1.2",
-    "mongodb-client-encryption": "^6.4.0",
+    "mongodb-client-encryption": "^6.5.0",
     "mongodb-legacy": "^6.1.3",
     "nyc": "^15.1.0",
-    "prettier": "^3.5.3",
-    "semver": "^7.7.0",
+    "prettier": "^3.6.2",
+    "semver": "^7.7.2",
     "sinon": "^18.0.1",
     "sinon-chai": "^3.7.0",
-    "snappy": "^7.2.2",
-    "socks": "^2.8.1",
+    "snappy": "^7.3.2",
+    "socks": "^2.8.7",
     "source-map-support": "^0.5.21",
     "ts-node": "^10.9.2",
-    "tsd": "^0.32.0",
+    "tsd": "^0.33.0",
     "typescript": "5.8.3",
     "typescript-cached-transpile": "^0.0.6",
     "v8-heapsnapshot": "^1.3.1",
